09_lesson/requirements.txt
SQLAlchemy==2.0.0
pytest==7.0.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
